tts-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
numpy
soundfile
librosa
boto3
requests
google-genai
